feat : 상품 조회 api를 최적화 했어요.

+++ src/main/kotlin/org/teamalilm/alilmbe/controller/ProductListController.kt
@@ -93,7 +93,7 @@ class ProductListController(
         val category: String,
         val price: Int,
         val productInfo: ProductInfo,
-        val waitingCount: Int,
+        val waitingCount: Long,
         val oldestCreationTime: Long
     )
